Remove obsolete windows rank helpers
