+++ affine/api/routers/__init__.py
@@ -9,7 +9,6 @@ from affine.api.routers.logs import router as logs_router
 from affine.api.routers.miners import router as miners_router
 from affine.api.routers.rank import router as rank_router
 from affine.api.routers.scores import router as scores_router
-from affine.api.routers.windows import router as windows_router
 
 __all__ = [
     "config_router",
@@ -17,5 +16,4 @@ __all__ = [
     "miners_router",
     "rank_router",
     "scores_router",
-    "windows_router",
 ]
